requirements.txt
python-socketio[client]
aiohttp
fastapi
uvicorn
sqlalchemy
pandas
numpy
pydantic
requests
